Fixed bad variable types and declarations

+++ MasterCan/WordCreator.cpp
@@ -11,7 +11,7 @@ transportObject::transportObject() {
 }
 
 // Concatenates all the data into a single string
-String transportObject::create(int packetCount, int nowTime, float pressure, float temperature,  float altitude, float velocity, float battVolt, byte softState, imu::Vector<3> accelData){
+String transportObject::create(unsigned long packetCount, unsigned long nowTime, float pressure, float temperature,  float altitude, float velocity, float battVolt, byte softState, imu::Vector<3> accelData){
   // Delimiter is ":"
   return String(packetCount)+":"+String(nowTime)+":"+String(pressure,1)+":"+String(temperature,1)+":"+String(altitude,1)+":"+String(velocity,1)+":"+String(battVolt,0)+":"+String(softState)+":"+String(accelData.x(),2)+":"+String(accelData.y(),2)+":"+String(accelData.z(),2);
